Finally fixed old bugs related to VBLANK interrupts (0xE2 and 0xF2 opcodes had wrong operand length)
Tetris and other games would freeze when interrupts were serviced. I found out that Tetris, on startup, loads a DMA routine into HRAM for later use. This uses a few instructions to load it from memory, including the 0xE2 opcode. This opcode had an incorrect operand length (1 instead of 0) which caused an INC C instruction to be skipped, therefore all the code was copied into the same address, and when a VBLANK IRQ was serviced, it called into HRAM which contained incorrect code, executed a bunch of NOPs, overflowed the pc, and kept executing NOPs until the next VBLANK IRQ.